feat: 통계 페이지 하단 리스트 컴포넌트 분리 및 출력
