+++ client/src/styles/LineChart.css
@@ -1,8 +1,18 @@
 .line-chart__container {
-	margin: 4rem auto;
-	background-color: var(--color-white);
+	margin: 0 auto;
 	max-width: 848px;
 	display: flex;
 	flex-direction: column;
 	align-items: center;
+}
+
+.line-chart__container canvas {
+	margin-bottom: 2rem;
+	border: 1px solid var(--color-gray-200);
+	box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1), 0px 4px 20px rgba(0, 0, 0, 0.1);
+	border-radius: 10px;
+}
+
+.line-chart__container ol {
+	width: 100%;
 }
